# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "To-Do List API",
    "version": "1.0.0",
    "description": "API para registro/login de usuários e gerenciamento de tarefas (CRUD + conclusão)."
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
    "/register": {
      "post": {
        "summary": "Registrar novo usuário",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": { "type": "string" },
                  "password": { "type": "string" }
                },
                "required": ["username", "password"]
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Usuário registrado" },
          "400": { "description": "Erro de registro" }
        }
      }
    },
    "/login": {
      "post": {
        "summary": "Login de usuário",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": { "type": "string" },
                  "password": { "type": "string" }
                },
                "required": ["username", "password"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Login realizado (retorna JWT)" },
          "400": { "description": "Erro de login" },
          "401": { "description": "Credenciais inválidas" }
        }
      }
    },
    "/tasks": {
      "post": {
        "summary": "Criar uma nova task",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": { "type": "string" },
                  "description": { "type": "string" },
                  "dueDate": { "type": "string", "format": "date" },
                  "priority": { "type": "string", "enum": ["low", "medium", "high"] }
                },
                "required": ["title", "dueDate"]
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Task criada" },
          "400": { "description": "Erro de validação" },
          "401": { "description": "Token ausente ou inválido" }
        }
      },
      "get": {
        "summary": "Listar todas as tasks do usuário autenticado",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "Lista de tasks" },
          "401": { "description": "Token ausente ou inválido" }
        }
      }
    },
    "/tasks/{id}": {
      "get": {
        "summary": "Obter detalhes de uma task",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "Task encontrada" },
          "404": { "description": "Task não encontrada" },
          "401": { "description": "Token ausente ou inválido" }
        }
      },
      "put": {
        "summary": "Atualizar uma task",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": { "type": "string" },
                  "description": { "type": "string" },
                  "dueDate": { "type": "string", "format": "date" },
                  "priority": { "type": "string", "enum": ["low", "medium", "high"] },
                  "status": { "type": "string", "enum": ["pending", "done"] }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Task atualizada" },
          "400": { "description": "Erro de validação" },
          "404": { "description": "Task não encontrada" },
          "401": { "description": "Token ausente ou inválido" }
        }
      },
      "delete": {
        "summary": "Excluir uma task",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "Task excluída com sucesso" },
          "404": { "description": "Task não encontrada" },
          "401": { "description": "Token ausente ou inválido" }
        }
      }
    },
    "/tasks/{id}/complete": {
      "post": {
        "summary": "Marcar uma task como concluída",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "Task marcada como concluída" },
          "400": { "description": "Task já estava concluída" },
          "404": { "description": "Task não encontrada" },
          "401": { "description": "Token ausente ou inválido" }
        }
      }
    }
  }
}